set up git ignore and no db in the repo??

+++ foodapp/macros.py
@@ -14,12 +14,15 @@ def calc_macros():
         size = input("Enter amount: ")
         if size == 'x':
             return None
-        ratio = int(size) / food_by_name.ss()
+        ratio = float(size) / food_by_name.ss()
         print(f"ratio: {ratio}") # testing
         altered_food = FoodItem()
         altered_food.set_info_from_tuple(food_by_name.get_tuple())
         altered_food.proportionalize(ratio)
         return altered_food
+
+
+
